chore: Moved SchemaDocuentScheduler to a context folder
