+++ integration/schema-language-server/language-server/src/main/java/ai/vespa/schemals/context/SchemaDocumentParser.java
@@ -1,4 +1,4 @@
-package ai.vespa.schemals.parser;
+package ai.vespa.schemals.context;
 
 import java.io.PrintStream;
